+++ app/Http/Controllers/ForumController.php
@@ -22,26 +22,37 @@ class ForumController extends Controller
 
     public function edit($id)
     {
+        
         $categories=Category::orderby('title', 'asc')->get();
         $post=Post::find($id);
+        if(auth()->user()->id == $post->user_id){
         return view('forum.edit')
         ->with('post', $post)
         ->with('categories', $categories);
+        }else{
+            return redirect('/');
+        }
     }
 
     public function create()
     {
-        $categories=Category::orderby('title', 'asc')->get();
-        return view('forum.create')->with('categories', $categories);
+        if(isset(auth()->user()->id)){
+            $categories=Category::orderby('title', 'asc')->get();
+            return view('forum.create')->with('categories', $categories);
+        } else{
+            return redirect('/');  
+        }
     }
 
     public function store(Request $request)
     {
-       $post=new Post;
-       $post->title=$request->title; 
-       $post->category_id=$request->category_id; 
-       $post->content=$request->content; 
-       $post->save();
+        if(isset(auth()->user()->id)){
+            $post=new Post;
+            $post->title=$request->title; 
+            $post->category_id=$request->category_id; 
+            $post->content=$request->content; 
+            $post->save();
+        } 
 
        $result=$request->all();
 
@@ -54,11 +65,13 @@ class ForumController extends Controller
 
     public function update(Request $request)
     {
-       $post=Post::find($request->post_id);
-       $post->title=$request->title; 
-       $post->category_id=$request->category_id; 
-       $post->content=$request->content; 
-       $post->save();
+        if(isset(auth()->user()->id)){
+            $post=Post::find($request->post_id);
+            $post->title=$request->title; 
+            $post->category_id=$request->category_id; 
+            $post->content=$request->content; 
+            $post->save();
+        }
 
        $result=$request->all();
 
@@ -84,8 +97,9 @@ class ForumController extends Controller
     {
         
         $post=Post::find($id);
-        $post->delete();
-
+        if(auth()->user()->id == $post->user_id){
+           $post->delete();
+        }
         return redirect('/');
     }
